refactor: improve naming for better consistency
- Use term 'attempt' instead of 'pass'
- Rename 'TestResultManager' and components to 'Repository' and merge with existing repository class
- Adjust naming in DIC
- Use term 'ParticipantResult' instead of 'TestResult'
