+++ components/ILIAS/Test/classes/class.ilObjTestGUI.php
@@ -200,7 +200,7 @@ class ilObjTestGUI extends ilObjectGUI implements ilCtrlBaseClassInterface, ilDe
         $this->export_factory = $local_dic['exportimport.factory'];
         $this->export_repository = $local_dic['exportimport.repository'];
         $this->participant_access_filter_factory = $local_dic['participant.access_filter.factory'];
-        $this->test_pass_result_repository = $local_dic['results.data.test_result_repository'];
+        $this->test_pass_result_repository = $local_dic['results.data.repository'];
         $this->toplist_repository = $local_dic['results.toplist.repository'];
 
         $ref_id = 0;
