+++ components/ILIAS/Test/tests/ilTestBaseTestCaseTrait.php
@@ -506,7 +506,7 @@ trait ilTestBaseTestCaseTrait
             => $this->createMock(
                 ILIAS\Test\Results\Toplist\TestTopListRepository::class
             );
-        $local_dic_mock['results.data.test_result_repository'] = fn(Pimple\Container $c)
+        $local_dic_mock['results.data.repository'] = fn(Pimple\Container $c)
             => $this->createMock(
                 \ILIAS\Test\Results\Data\Repository::class
             );
